Moving pidap files to top level under album_streamer

+++ buildroot_pidap/board/PIDAP/post-build.sh
@@ -6,11 +6,18 @@
 
 set -e
 
-# Strip release-image network/cron init scripts; openssh can still be
-# installed but we don't need eth0 to be configured.
-rm -f "${TARGET_DIR}/etc/init.d/S39wait-eth0"
-rm -f "${TARGET_DIR}/etc/init.d/S40network"
-rm -f "${TARGET_DIR}/etc/init.d/S41static-eth"
+# Release build has no openssh => remove the sshd init script.
+if [ ! -f "${TARGET_DIR}/usr/sbin/sshd" ]; then
+    rm -f "${TARGET_DIR}/etc/init.d/S50sshd"
+fi
+
+# Release build has no ifupdown => remove the network init scripts.
+if [ ! -x "${TARGET_DIR}/usr/sbin/ifup" ]; then
+    rm -f "${TARGET_DIR}/etc/init.d/S39wait-eth0"
+    rm -f "${TARGET_DIR}/etc/init.d/S40network"
+    rm -f "${TARGET_DIR}/etc/init.d/S41static-eth"
+    rm -f "${TARGET_DIR}/etc/network/interfaces"
+fi
+
+# Remove cron in all builds (not used).
 rm -f "${TARGET_DIR}/etc/init.d/S50crond"
-rm -f "${TARGET_DIR}/etc/init.d/S50sshd"
-rm -f "${TARGET_DIR}/etc/network/interfaces"
